Added first version of context
Created file todoContext and shared this context witt all components throught App and TodoContainer
